requirements.txt
# Core Trading Libraries
ccxt>=4.1.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.0

# Technical Analysis
ta>=0.10.2
pandas-ta>=0.3.14b

# Database
h5py>=3.8.0

# Optimization & Machine Learning
scikit-optimize>=0.9.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Progress & Logging
tqdm>=4.65.0
colorama>=0.4.6
tabulate>=0.9.0

# HTTP & Networking
requests>=2.31.0
urllib3>=2.0.0
websocket-client>=1.6.0
aiohttp>=3.8.0

# Data Processing
python-dateutil>=2.8.2
pytz>=2023.3
jsonschema>=4.17.0

# Performance Optimization
numba>=0.57.0
bottleneck>=1.3.7
cython>=3.0.0

# Testing & Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: Notifications
python-telegram-bot>=20.0
# discord.py>=2.3.0
# twilio>=8.5.0

# Optional: Web Interface
# flask>=2.3.0
# flask-cors>=4.0.0
# dash>=2.10.0
# streamlit>=1.25.0

# Optional: Advanced ML
# optuna>=3.3.0
# hyperopt>=0.2.7
# ray[tune]>=2.6.0

# Optional: Database Alternatives
# pymongo>=4.4.0
# redis>=4.6.0
# influxdb-client>=1.36.0

# Optional: Cloud Deployment
# boto3>=1.28.0
# google-cloud-storage>=2.10.0
# azure-storage-blob>=12.17.0